This commit solves the following issues: 1. Avoid the use of center tag 2. Create main primary color variable 3. Make use of the English language 4. Comment badge bad positioning 5. Make use flex-box for layout 6. Create a main container 7. Make the viewport responsive ready 8. Improve PUG syntax
Have in mind each issue was supposed to have a different branch and
commit, but I was in doubt if that was the way to proceed at that point
and while I was waiting for the answer I started to solve these issues
in one branch.
Miguel Henao allowed me to do one commit only to avoid reverting and
making the changes one by one again.